requirements.txt
pytest~=9.0.3
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
pytest-base-url
allure-pytest
requests~=2.33.1
jsonpath~=0.82.2
pyyaml~=6.0.3
Cerberus~=1.3.8